fix warnings

+++ src/debug/structures_size.c
@@ -23,15 +23,15 @@
 void _starpu_debug_display_structures_size(void)
 {
 	fprintf(stderr, "struct starpu_task\t\t%d bytes\t(%x)\n",
-			sizeof(struct starpu_task), sizeof(struct starpu_task));
+			(unsigned) sizeof(struct starpu_task), (unsigned) sizeof(struct starpu_task));
 	fprintf(stderr, "struct starpu_job_s\t\t%d bytes\t(%x)\n",
-			sizeof(struct starpu_job_s), sizeof(struct starpu_job_s));
+			(unsigned) sizeof(struct starpu_job_s), (unsigned) sizeof(struct starpu_job_s));
 	fprintf(stderr, "struct starpu_data_state_t\t%d bytes\t(%x)\n",
-			sizeof(struct starpu_data_state_t), sizeof(struct starpu_data_state_t));
+			(unsigned) sizeof(struct starpu_data_state_t), (unsigned) sizeof(struct starpu_data_state_t));
 	fprintf(stderr, "struct starpu_tag_s\t\t%d bytes\t(%x)\n",
-			sizeof(struct starpu_tag_s), sizeof(struct starpu_tag_s));
+			(unsigned) sizeof(struct starpu_tag_s), (unsigned) sizeof(struct starpu_tag_s));
 	fprintf(stderr, "struct starpu_cg_s\t\t%d bytes\t(%x)\n",
-			sizeof(struct starpu_cg_s), sizeof(struct starpu_cg_s));
+			(unsigned) sizeof(struct starpu_cg_s), (unsigned) sizeof(struct starpu_cg_s));
 	fprintf(stderr, "struct starpu_worker_s\t\t%d bytes\t(%x)\n",
-			sizeof(struct starpu_worker_s), sizeof(struct starpu_worker_s));
+			(unsigned) sizeof(struct starpu_worker_s), (unsigned) sizeof(struct starpu_worker_s));
 }
